// services/src/inventory/model/InventoryEventName.ts
export enum InventoryEventName {
  SKU_RESTOCKED_EVENT = 'SKU_RESTOCKED_EVENT',
  ORDER_CREATED_EVENT = 'ORDER_CREATED_EVENT',
  ORDER_STOCK_ALLOCATED_EVENT = 'ORDER_STOCK_ALLOCATED_EVENT',
  ORDER_STOCK_DEPLETED_EVENT = 'ORDER_STOCK_DEPLETED_EVENT',
  ORDER_PAYMENT_REJECTED_EVENT = 'ORDER_PAYMENT_REJECTED_EVENT',
  ORDER_CANCELED_EVENT = 'ORDER_CANCELED_EVENT',
}